Allow overriding entity cache parameters
We're putting together a site where the language negotiation is done via an HTTP `Accept-Language:` header, but this resulted in cache conflicts since there were no parameters in the URL on which the differing language values could be cached. Unfortunately, with the way `viewEntity()` is structured we had to copy/paste the entire function just to change the cache tagging -- it seems better to allow this to be overridden at the plugin layer so I did a small re-factoring to support this.