[FEATURE] (Entity) Allow a different ID field
This pull request allows you to specify a field to be used as the
primary ID. That means that all of the entity based resources need
to have the `id` field present. The `id` field is inherited from
the abstract base class ResourceEntity, but it can be swapped by
another one. If you don't want the `id` field present in your
resource input & output you can use the `'methods' => array()`
trick in the field definition. If you specify an `idField` in the
resource annotation, that field will be
used as the loadByIdField by default. This allows you to stop using
entity ids as your API ids and use UUIDs (maybe generated by the
uuid Drupal module).