[BUGFIX] Fix includes and fields for nested
Pass the request down to the resource fields so we can pass relevant
query string parameters to compound documents. The parameters are
detected by stripping the public name + a dot from the beginnining
of the parameter. So if 'related.images.uri' is passed to the top
request, 'images.uri' will be passed to the ResourceFieldResource
for 'related'. The resource field for 'images' will recieve 'uri'
in the generated request object to compound the document.