[BUGFIX] Make the relationships more generic
Remove the case by case handling (entityreference, taxonomy term,
file and uid) in favor of a more generic one leveraging the metadata
in the referenced resource definition. To get the referenced ID when
there is a 'referencedIdProperty' we only need the entity type and
the bundle. Both of these are present in the destination resource
definition.

Additionally use a similar technique to join entity tables in our
custom EFQ, when the referenced entity ID is held in a property
(instead of a field).