In-memory caching of authentication plugins.
When there are a lot of sub-entities returned from a rest call
the getDefinitions method gets called often and each time
it hits the database to return the cached authentication plugins.
This could result in dozens or hundreds of redundant database
calls per request.

Adding a memory cache of the definitions ensures there are no
redundant database calls. This cut the TTFB of our in-the-wild
service by more than half.