Improvement.
This change makes it possible to set an option for creating a subdirectory when saving a file, & it also adds the ability to optionally places files in a dated (Y-m-d) subdirectory within your named subdirectory.

This mimics the ability to set subdirectories & dated subdirectories when cteating a file field on a content type in Drupal core. For instance, you can add a file field & in the field settings you can put something like the following in the 'File directory' field: `avatars/[current-date:custom:Y]-[current-date:custom:m]-[current-date:custom:d`.