Add ability to retrieve resource name from the resource manager
The ResourceManager class provides methods to get version information and even create a new
instance of a plugin based on the current request. However, the methods that help in getting
the resource id (such as ::getMenuItem and getPageArguments) are protected.

Since we are already able to get information on version, we should also be able to get the
resource id using the resource manager class. This commit adds a method to do that.

This blocks issue #939 where we need to get the resource plugin definition from an
authentication plugin. Calling negotiate or trying to retrieve menu item information via
menu_get_item results in an recursive call (infinite).