Add OAuth2 authentication provider
This change depends on the changes in PR #940. It uses the new method introduced in that PR
to identify the resource being requested and get the plugin definition to read the oauth2
server and scope to use for authorizing the request.

This would probably need to be changed once #940 is merged or a better method to access the
request being processed is identifed (as mentioned in #939). I am creating this PR in hope
of clarifying the changes in PR #940.