[FEATURE] Allow getting resource name from resource manager (#940)
The ResourceManager class provides methods to get version information and even create a new
instance of a plugin based on the current request. However, the methods that help in getting
the resource id (such as ::getMenuItem and getPageArguments) are protected.

Since we are already able to get information on version, we should also be able to get the
resource id using the resource manager class. This commit adds a method to do that.