Use arrays to define OAuth2 information in the plugin definition
Currently, we use two pieces of information for OAuth2 authentication - server and scope. We define them as separate items in the plugin definition. There might be more
items required in the future (one of which could be support for client credentials in #944). It might be cleaner to change all the definitions to an array.